ChangesInBlog

+++ controllers/blogController.js
@@ -7,30 +7,47 @@ const postBlog = async (req, res) => {
     title,
     user: req.user.userId,
   });
-  console.log(blog);
+  //console.log(blog);
   await blog.save();
 
   res.status(201).json({ msg: "Blog Created Successfully" });
 };
 
 const updateBlog = async (req, res) => {
-  const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, req.body, {
-    new: true,
-  });
-  if (!req.params.id) {
-    return res.status(400).json({ error: "No Blog" });
+  const blog = await Blog.findById(req.params.id);
+
+  // console.log(blog.user);
+  const currentuser = req.user.userId;
+  //console.log(currentuser);
+  if (blog.user == currentuser) {
+    const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, req.body, {
+      new: true,
+    });
+
+    if (!req.params.id) {
+      return res.status(400).json({ error: "No Blog" });
+    }
+    res.send({ msg: "Updated Successfully", updateBlog });
+    await updatedBlog.save();
+  } else {
+    res.status(400).json({ error: "No permission to update" });
   }
-  res.send(updatedBlog);
-  await updatedBlog.save();
 };
 
 const deleteBlog = async (req, res) => {
   try {
-    const deletedblog = await Blog.findByIdAndDelete(req.params.id);
-    if (!req.params.id) {
-      return res.status(400).json({ error: "No Blog" });
+    const blog = await Blog.findById(req.params.id);
+    const currentuser = req.user.userId;
+
+    if (blog.user == currentuser) {
+      const deletedblog = await Blog.findByIdAndDelete(req.params.id);
+      if (!req.params.id) {
+        return res.status(400).json({ error: "No Blog" });
+      }
+      res.send({ msg: "Deleted Successfully", deletedblog });
+    } else {
+      res.status(400).json({ error: "No permission to delete" });
     }
-    res.send(deletedblog);
   } catch (error) {
     res.status(500).json({ error: "Internal Server Error" });
   }
